Allow passing WebSocket credentials with Sec-WebSocket-Protocol header
This is a work-around to the problem that the JavaScript WebSocket
implementation doesn't support sending custom headers, such as the
Authorization header.

Although this is not optimal, it is still considered secure due to
the fact that the credentials are sent in the headers and will
therefore be encrypted and are unlikely to be logged on any servers.
And since Sec-WebSocket-Protocol is specified in the RFC 6455
standard (https://tools.ietf.org/html/rfc6455#section-11.3.4), it is
very unlikely that this will change any time soon and will probably
remain secure over many different implementations.