Generate a unique mnemonic seed for each user node
This seed is normally used by lnd's internal wallet, but that wallet
is not used in pine-lnd since it's been moved to the Pine app.
However, the seed is still used to give the node its pubkey that is
used for identification, thus giving each user node its own unique
pubkey.