Increment key index for deriveNextKey and getRevocationRootKey
Increment key index so that a new key is always returned. This index
is stored on the server side instead of client side to make sure that
it's not lost in case the user has to restore their account.

This is done by annotating the request to the Pine app with the current
key index for the relevant user and method before the request is proxied.

For `deriveNextKey` and `getRevocationRootKey`.